move FEATURE_COLLECTION_GEORESOURCE_ID from FeatureCollectionPlugin to domain/geoResources
